pyproject.toml
[build-system]
requires = ['setuptools>=61']
build-backend = "setuptools.build_meta"

[project]
name = "sshcheck"
authors = [
    {name="Johnny Birchett", email="lol@ok.wutevs"},
]
description = "A small utility for checking ssh hosts against a YAML Key Exchange security policy"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta"
]
version = "0.0.5"
dependencies = [
    'paramiko', 'typer[all]', 'pyyaml',
    'importlib-metadata; python_version<"3.8"'
]
keywords = ['ssh', 'kex', 'configuration']

[project.optional-dependencies]
img = ['cairosvg']




[project.urls]
Homepage = "https://github.com/XanaDublaKublaConch/sshcheck"
Repository = "https://github.com/XanaDublaKublaConch/sshcheck"

[project.scripts]
sshcheck = "sshcheck.cli:run_cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
sshcheck = ["policy.yml"]
